+++ src/Encoder/SimpleIDEncoder.php
@@ -11,10 +11,11 @@
 namespace Ynlo\GraphQLBundle\Encoder;
 
 use Doctrine\Bundle\DoctrineBundle\Registry;
-use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManagerInterface;
 use Ynlo\GraphQLBundle\Definition\Registry\DefinitionRegistry;
+use Ynlo\GraphQLBundle\Error\NodeNotFoundException;
 use Ynlo\GraphQLBundle\Model\NodeInterface;
+use Ynlo\GraphQLBundle\Util\TypeUtil;
 
 class SimpleIDEncoder implements IDEncoderInterface
 {
@@ -42,8 +43,7 @@ class SimpleIDEncoder implements IDEncoderInterface
      */
     public function encode(NodeInterface $node): ?string
     {
-        $class = ClassUtils::getClass($node);
-        $nodeType = $this->definitionRegistry->getEndpoint()->getTypeForClass($class);
+        $nodeType = TypeUtil::resolveNodeType($this->definitionRegistry->getEndpoint(), $node);
 
         return sprintf('%s%s%s', $nodeType, self::DIVIDER, $node->getId());
     }
@@ -59,7 +59,17 @@ class SimpleIDEncoder implements IDEncoderInterface
             $class = $this->definitionRegistry->getEndpoint()->getClassForType($nodeType);
             $manager = $this->doctrine->getManager();
             if ($manager instanceof EntityManagerInterface) {
-                return $manager->getReference($class, $databaseId);
+                $reference = $manager->getReference($class, $databaseId);
+                $resolvedType = TypeUtil::resolveNodeType($this->definitionRegistry->getEndpoint(), $reference);
+
+                //compare the given type encoded in the globalID with the type resolved by the object instance
+                //This is important to avoid get a node using different type when use the same entity class
+                //e.g. 'AdminUser:1' => resolve to type 'AdminUser' and should not be possible get using 'CommonUser:1' as globalId
+                if ($resolvedType !== $nodeType) {
+                    return null;
+                }
+
+                return $reference;
             }
 
             throw new \UnexpectedValueException('Not supported doctrine manager.');